Fixed fee printing bug, Added 0.5,3 seconds wait ind bank actions

+++ hw2/Account.h
@@ -23,7 +23,7 @@ public:
     void BecomeVip();
     void Deposit(int );
     int Withdrawal(int );   // 0 if succescul   1 if not enough money
-    int GiveInterest (double percent);
+    int GiveInterest (int percent);
 
 private:
     int _accountNumber;
